requirements.txt
pandas
numpy==1.23.5
python-binance
python-telegram-bot
prometheus-client
pandas-ta
pybit
